+++ packages/suite/src/hooks/wallet/useCardanoStaking.ts
@@ -1,11 +1,9 @@
 import { useState, useCallback, useMemo } from 'react';
+
 import { ActionAvailability, CardanoStaking } from 'src/types/wallet/cardanoStaking';
 import { SUITE } from 'src/actions/suite/constants';
-import trezorConnect, { PROTO } from '@trezor/connect';
 import { useActions, useSelector } from 'src/hooks/suite';
-import { notificationsActions } from '@suite-common/toast-notifications';
 import * as cardanoStakingActions from 'src/actions/wallet/cardanoStakingActions';
-import { isTestnet } from '@suite-common/wallet-utils';
 import {
     getStakingPath,
     getProtocolMagic,
@@ -17,9 +15,12 @@ import {
     getStakePoolForDelegation,
     getTtl,
     loadCardanoLib,
-    getDerivationType,
 } from 'src/utils/wallet/cardanoUtils';
 import { AppState } from 'src/types/suite';
+
+import { isTestnet, getDerivationType } from '@suite-common/wallet-utils';
+import { notificationsActions } from '@suite-common/toast-notifications';
+import trezorConnect, { PROTO } from '@trezor/connect';
 import { addFakePendingCardanoTxThunk } from '@suite-common/wallet-core';
 
 const getDeviceAvailability = (
