+++ suite-common/wallet-utils/src/index.ts
@@ -13,5 +13,6 @@ export * from './exportTransactions';
 export * from './tokenUtils';
 export * from './localizeNumber';
 export * from './backend';
+export * from './cardanoUtils';
 
 export { analyzeTransactions as analyzeTransactionsFixtures } from './__fixtures__/transactionUtils';
